arm64/sysreg: Convert SPE registers to automatic generation
Convert all the SPE register defines to automatic generation. No
functional changes.

New registers and fields for SPEv1.2 are added with the conversion.

Some of the PMBSR MSS field defines are kept as the automatic generation
has no way to create multiple names for the same register bits. The
meaning of the MSS field depends on other bits.

Signed-off-by: Rob Herring <robh@kernel.org>
Message-Id: <20221019191126.GH1jyFkqAORpc5j0fScbQ3Tibc6yPQiMomO9Y5ucAek@z>