arm64: Drop SYS_ from SPE register defines
We currently have a non-standard SYS_ prefix in the constants generated
for the SPE register bitfields. Drop this in preparation for automatic
register definition generation.

The SPE mask defines were unshifted, and the SPE register field
enumerations were shifted. The autogenerated defines are the opposite,
so make the necessary adjustments.

No functional changes.

Tested-by: James Clark <james.clark@arm.com>
Signed-off-by: Rob Herring <robh@kernel.org>
Message-Id: <20221104155502.mcxqZIXXg-5AAuqPxXjDLb57FKDJLpnieMEg2t7WKKM@z>