KVM: arm64: Handle CCSIDR associativity mismatches
CCSIDR associativity mismatches among the physical CPUs causes a vCPU
see inconsistent values when it is migrated among physical CPUs.

It also makes QEMU fail restoring the vCPU registers because QEMU saves
and restores all of the registers including CCSIDRs, and if the vCPU
migrated among physical CPUs between saving and restoring, it tries to
restore CCSIDR values that mismatch with the current physical CPU, which
causes EFAULT.

Trap CCSIDRs if there are CCSIDR value msimatches, and override the
associativity bits when handling the trap.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>
Reported-by: Akihiko Odaki <akihiko.odaki@gmail.com>
Message-Id: <20221201104914.yxibloeNvRvU9HNpqTojNqFHoaokG8PGM8DcmGqw-rc@z>