KVM: arm64: Always mask CCSIDR associativity bits
M2 MacBook Air has mismatched CCSIDR associativity bits among physical
CPUs, which makes the bits a KVM vCPU sees inconsistent when migrating
among them.

While it is possible to detect CCSIDR associativity bit mismatches and
mask them with that condition, it requires mismatch detection and
increases complexity. Instead, always mask the CCSIDR associativity bits
to keep the code simple.

Also, allow the userspace to overwrite the bits with arbitrary values so
that it can restore a vCPU state saved with an older kernel.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>
Suggested-by: Marc Zyngier <maz@kernel.org>
Message-Id: <20221202091856.aLU6AG8_pPny4eoOQXZvsEaap6J03Z21jUQZ5tJPOeg@z>