KVM: arm64: Remove use of ARM64_FEATURE_MASK()
The KVM code makes extensive use of ARM64_FEATURE_MASK() to generate a
mask for fields in the ID registers. This macro has the assumption that
all feature fields are 4 bits wide but the architecture has evolved to
add fields with other widths, such as the 1 bit fields in ID_AA64SMFR0_EL1,
so we need to adjust the

We could fix this by making ARM64_FEATURE_MASK() use the generated macros
that we have now but since one of these is a direct _MASK constant the
result is something that's more verbose and less direct than just updating
the users to directly use the generated mask macros, writing

	#define ARM64_FEATURE_MASK(x)	(x##_MASK)

obviously looks redundant and if we look at the users updating them turns

	val &= ~ARM64_FEATURE_MASK(ID_AA64PFR0_EL1_CSV3);

into the more direct

	val &= ~ID_AA64PFR0_EL1_CSV3_MASK;

rather than updating the macro just remove the users. This is a relatively
large code change but very mechanical.

No functional change.

Signed-off-by: Mark Brown <broonie@kernel.org>
Message-Id: <20221221180610.2fhx-bJAjzOeOdy0XtMwx5q4hJSP_jns7PD7P2O9d8Q@z>