KVM: arm64: Reuse fields of sys_reg_desc for idreg
Since reset() and val are not used for idreg in sys_reg_desc, they would
be used with other purposes for idregs.
The callback reset() would be used to return KVM sanitised id register
values. The u64 val would be used as mask for writable fields in idregs.
Only bits with 1 in val are writable from userspace.

Signed-off-by: Jing Zhang <jingzhangos@google.com>
Message-Id: <20230404035344.4043856-6-jingzhangos@google.com>