KVM: arm64: selftests: Delete core_reg_fixup
core_reg_fixup() complicates sharing the get-reg-list test with
other architectures. Rather than work at keeping it, with plenty
of #ifdeffery, just delete it, as it's unlikely to test a kernel
based on anything older than v5.2 with the get-reg-list test,
which is a test meant to check for regressions in new kernels.
(And, an older version of the test can still be used for older
kernels if necessary.)

Signed-off-by: Andrew Jones <ajones@ventanamicro.com>
Message-Id: <2cdacf18888f9cfe020d218adc94e4b3b657d3e8.1684999824.git.haibo1.xu@intel.com>