KVM: riscv: selftests: Make check_supported arch specific
check_supported() was used to verify whether a feature/extension was
supported in a guest in the get-reg-list test. Currently this info
can be retrieved through the KVM_CAP_ARM_* API in aarch64, but in
riscv, this info was only exposed through the KVM_GET_ONE_REG on
KVM_REG_RISCV_ISA_EXT pseudo registers.

Signed-off-by: Haibo Xu <haibo1.xu@intel.com>
Message-Id: <26dea518fc5e8da51e61db279d175364bfecd009.1684999824.git.haibo1.xu@intel.com>