KVM: arm64: Keep need_db always true in vgic_v4_put() when emulating WFI
When enabled GICv4.1 on Kunpeng 920 platform with 6.4 kernel (preemptible
kernel), running a vm with 128 vcpus on 64 pcpu, sometimes vm is not booted
successfully, and we find there is a situation that doorbell interrupt will
not occur event if there is a pending interrupt.
In function kvm_vcpu_wfi(), the parameter of need_db is true for function
vgic_v4_put() which wants to tell GICv4 that we need doorbells to be
signalled if there is a pending interrupt.
But if there is a preemption before kvm_vcpu_halt (after preempt_enable()),
it will change the value of vpe->resident, which will make vgic_v4_put()
is called with need_db = false When calling function schedule().
Then after calling schedule(), doorbell interrupt will not occur even
if there is a pending interrupt.
We need to keep need_db always true for emulate Wait-For-Interrupt
behavior, so set need_db true in vgic_v4_put() if it is in the process of
emulating WFI.

kvm_vcpu_wfi
    preempt_disable
    ...
    vgic_v4_put(vcpu, true)
	vpe->resident = 0
	need_db = 1
    preempt_enable
Tested-by: Xiang Chen <chenxiang66@hisilicon.com>
------------------------------------------> preemption occur
    kvm_sched_out
	vgic_v4_put(vcpu, false)
	    vpe->resident == 0 -> return 0 (do nothing)
	    ...
<------------------------------------------ back
    kvm_sched_in
	vgic_v4_load
	    vpe->resident  = 1
	    ...
------------------------------------------- Continue
    kvm_vcpu_halt
	set vcpu thread INTERRUPTIBLE
	schedule() ->
	    kvm_sched_out
		vgic_v4_put(vcpu, false)
		    need_db = 0 (vcpu thread is not scheduled and doorbell interrupt
		    will not signalled even if there is a pending interrupt)
		    vpe->resident = 0

Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Message-Id: <1688720145-37923-1-git-send-email-chenxiang66@hisilicon.com>