KVM: arm64: Configure HCRX_EL2 dynamically
At the moment the HCRX_EL2 system register is always initialized to
HCRX_GUEST_FLAGS when running a guest. Instead, choose the configuration
at vcpu reset time and save it in the vcpu struct, similarly to how
HCR_EL2 is set up. This will be needed in a subsequent change to
configure the register based on CPU features detected at runtime.

Signed-off-by: Kristina Martsenko <kristina.martsenko@arm.com>
Message-Id: <20230915124840.474888-2-kristina.martsenko@arm.com>