KVM: arm64: Use folio for THP adjustment
Now transparent huge pages use folios, we can check efficiently if a
page is mapped by a block by checking the folio size. This is saving a
page table walk.

It is safe to read the folio in this path. We've just increased its
refcount (GUP from __gfn_to_pfn_memslot()). This will prevent attempts
of splitting the huge page.

Signed-off-by: Vincent Donnefort <vdonnefort@google.com>

Message-Id: <20230926181932.1650692-3-vdonnefort@google.com>