arm64: KVM: Add interface to set guest value for TRFCR register
Add an interface for the Coresight driver to use to set the value of the
TRFCR register for the guest. This register controls the exclude
settings for trace at different exception levels, and is used to
honor the exclude_host and exclude_guest parameters from the Perf
session. This will be used to later write TRFCR_EL1 on nVHE at guest
switch. For VHE, TRFCR_EL1 is written immediately. Because guest writes
to the register are trapped, the value will persist and can't be
modified.

The settings must be copied to the vCPU before each run in the same
way that PMU events are because the per-cpu struct isn't accessible in
protected mode.

Now that both guest and host values are saved, rename trfcr_el1 to
host_trfcr_el1 to make it clear that's the value that should be restored
on return to the host.

Reviewed-by: Mark Brown <broonie@kernel.org> (sysreg)
Signed-off-by: James Clark <james.clark@arm.com>
Message-Id: <20230928151611.3042443-4-james.clark@arm.com>