arm: perf: Include threshold control fields valid in PMEVTYPER mask
FEAT_PMUv3_TH (Armv8.8) adds two new fields to PMEVTYPER, so include
them in the mask. These aren't writable on 32 bit kernels as they are in
the high part of the register, so split the mask definition to the asm
files for each platform.

Despite not being used on aarch32, TH and TC macros are added to the
shared header file, because will be used in arm_pmuv3.c which is
compiled for both platforms.

Signed-off-by: James Clark <james.clark@arm.com>
Message-Id: <20231025095710.1559601-2-james.clark@arm.com>